reorganize session tests, update comments
